chore: move pointer logic to a new partial
